--- src/main/java/io/github/teamwaff1e/waffle/global/dto/RequestDto.java
@@ -1,4 +1,0 @@
-package io.github.teamwaff1e.waffle.global.dto;
-
-public interface RequestDto {
-}
