+++ src/main/java/io/github/teamwaff1e/waffle/global/dto/request/RequestDto.java
@@ -1,0 +1,4 @@
+package io.github.teamwaff1e.waffle.global.dto.request;
+
+public interface RequestDto {
+}
