+++ src/main/java/io/github/teamwaff1e/waffle/global/dto/response/ResponseDto.java
@@ -1,0 +1,4 @@
+package io.github.teamwaff1e.waffle.global.dto.response;
+
+public interface ResponseDto {
+}
